ref: split GithubHandler into two classes
